Remove some unused variables.
The code now compiles mostly cleanly with -W -Wall -Wextra.
I've kept the various unused colors (crown_yellow, for example)
in case I ever want to redo that code someday without looking up
the color values anew.

No change in behavior.

+++ input.cc
@@ -64,7 +64,7 @@ struct Pixel {
 };
 
 struct PixelRegion {
-    int pop;
+    size_t pop;
     unsigned char color[3];
     std::vector<Pixel> pixels;
